Added desktop files and adjusted Makefile to install them. Updated modules to configure themselves when engaged.  For some reason they were not being configured until explicitely called by user.

+++ src/autopopups.d
@@ -131,7 +131,7 @@ class AUTO_POP_UPS
 
     Window                      mTipsWin;
     DATA_STORE[MAX_TIP_DEPTH]   mTipsStore;
-    size_t                      mTipsIndex;
+    int 	                    mTipsIndex;
     TreeView                    mTipsView;
